+++ cli/src/pcluster/templates/cw_dashboard_builder.py
@@ -752,7 +752,10 @@ class CWDashboardConstruct(Construct):
                     self._new_cw_log_widget(
                         title="system-messages",
                         conditions=[
-                            Condition(["alinux2", "alinux2023", "rhel8", "rocky8", "rhel9", "rocky9"], base_os)
+                            Condition(
+                                ["alinux2", "alinux2023", "rhel8", "rocky8", "rhel9", "rocky9", "almalinux8"],
+                                base_os,
+                            )
                         ],
                         filters=[self._new_filter(pattern=f"{head_private_ip}.*system-messages")],
                     ),