+++ cli/tests/pcluster/validators/test_cluster_validators.py
@@ -3663,6 +3663,7 @@ def test_multi_network_interfaces_instances_validator(
         ("ami-000000000000", "rocky8", None, None),
         ("ami-000000000000", "alinux2", None, None),
         ("ami-000000000000", "rhel8", None, None),
+        ("ami-000000000000", "almalinux8", None, None),
         ("ami-000000000000", "ubuntu2204", None, None),
         ("ami-000000000000", "ubuntu2404", None, None),
         (
@@ -3675,6 +3676,15 @@ def test_multi_network_interfaces_instances_validator(
             FailureLevel.ERROR,
         ),
         (None, "alinux2", None, None),
+        (
+            None,
+            "almalinux8",
+            "ParallelCluster has no official AMI for almalinux8. "
+            "Please build your own AMI using pcluster build-image command, "
+            "as explained in the documentation: "
+            "https://docs.aws.amazon.com/parallelcluster/latest/ug/building-custom-ami-v3.html",
+            FailureLevel.ERROR,
+        ),
         (None, "rhel8", None, None),
         (None, "ubuntu2204", None, None),
         (None, "ubuntu2404", None, None),